+++ config/site.nix
@@ -2,6 +2,7 @@
 # Host identity and primary user settings
 { ... }:
 {
+  hostSystem = "x86_64-linux";
   hostname = "nixoa";
   timezone = "UTC";
   stateVersion = "25.11"; # Don't change this after initial installation