+++ modules/aspects/nixoa-user.nix
@@ -1,4 +1,0 @@
-{ ... }:
-{
-  den.aspects.nixoaUser.homeManager = import ../_homeManager/profile;
-}