+++ modules/aspects/user.nix
@@ -1,0 +1,4 @@
+{ vars, ... }:
+{
+  den.aspects.${vars.username}.homeManager = import ../home;
+}