+++ modules/outputs/dev-shell.nix
@@ -5,7 +5,7 @@
   ...
 }:
 let
-  system = "x86_64-linux";
+  system = vars.hostSystem;
   pkgs = inputs.nixpkgs.legacyPackages.${system};
 in
 {